fixed memory leaks

+++ xmldb_lib/Hashtable.h
@@ -9,7 +9,7 @@ struct Hashtable{
     struct VectorInt* lists;
 };
 void free_hashtable(struct Hashtable* hashtable);
-void init_hashtable(struct Hashtable* hashtable ,int size);
+void init_hashtable(struct Hashtable* hashtable ,size_t size);
 void insertid_intohashtable(struct Hashtable* HashTB,int hashno ,int  nodeId,int reference_linenotoinsert,struct VectorInt *nodeNoToLineno) ;
 
 void removeid_fromhashtable(struct Hashtable* HashTB,int hashno ,int  nodeId,struct VectorInt *nodeNoToLineno);
